requirements.txt
transformers>=4.36.0
datasets>=2.14.0
peft>=0.6.0
torch>=2.0.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
trl>=0.7.2
wandb>=0.15.8
evaluate>=0.4.0
scipy>=1.11.0
scikit-learn>=1.3.0
PyYAML>=6.0.1 